requirements.txt
httpx
asyncio
concurrent.futures
rich
inquirerpy
pyyaml
nest_asyncio
